r18307: fixed a warning (This used to be commit cf24d46ea3c2dca4425121642400bef88f218a9e)

+++ source4/torture/basic/base.c
@@ -766,7 +766,7 @@ static BOOL run_deferopen(struct smbcli_state *cli, int dummy)
 			return False;
 		}
 
-		printf("pid %u open %d\n", getpid(), i);
+		printf("pid %u open %d\n", (unsigned)getpid(), i);
 
 		sleep(10);
 		i++;
